chore: move addExtendedOptions to evm/cli-utils
